Placed RSS into its plugin file
