+++ superform/superform/__init__.py
@@ -7,7 +7,6 @@ import superform.plugins
 from superform.authentication import authentication_page
 from superform.authorizations import authorizations_page
 from superform.edit import edit_page
-from superform.rss import rss_page
 from superform.stats import stats_page
 from superform.channels import channels_page
 from superform.delete import delete_page
@@ -17,6 +16,7 @@ from superform.search import search_page
 from superform.publishings import pub_page
 from superform.users import get_moderate_channels_for_user, is_moderator
 from superform.plugins.pdf import export
+from superform.plugins.rss import rss_page
 from superform.plugins._linkedin_callback import linkedin_page
 from superform.plugins._facebook_callback import facebook_page
